Remove failure-based implementation
